Authapi/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
firebase-admin>=6.2.0
pyrebase4>=4.8.0
pycryptodome>=3.19.0
requests>=2.27.1,<3.0.0
requests-toolbelt>=0.7.1,<1.0
